rename one of two getAssetDetail methods
