+++ kit/dapp/src/components/blocks/evm-address/evm-address.tsx
@@ -10,8 +10,8 @@ import {
 } from "@/components/ui/hover-card";
 import { Link } from "@/i18n/routing";
 import { getBlockExplorerAddressUrl } from "@/lib/block-explorer";
-import type { Asset } from "@/lib/queries/asset/asset-fragment";
 import { getAssetSearch } from "@/lib/queries/asset/asset-search";
+import type { AssetUsers } from "@/lib/queries/asset/asset-users-fragment";
 import type { User } from "@/lib/queries/user/user-fragment";
 import { getUserSearch } from "@/lib/queries/user/user-search";
 import { shortHex } from "@/lib/utils/hex";
@@ -79,7 +79,7 @@ export function EvmAddress({
 }: EvmAddressProps) {
   // State for user and asset data
   const [user, setUser] = useState<User | null>(null);
-  const [asset, setAsset] = useState<Asset | null>(null);
+  const [asset, setAsset] = useState<AssetUsers | null>(null);
 
   // Effect to fetch user and asset data
   useEffect(() => {
